+++ iftttie/templates/__init__.py
@@ -3,6 +3,7 @@ from __future__ import annotations
 from datetime import datetime, timedelta
 
 import aiohttp_jinja2
+import pprintpp
 from aiohttp.web_app import Application
 from jinja2 import PackageLoader, select_autoescape
 
@@ -14,6 +15,7 @@ def setup(app: Application):
     env.globals['Unit'] = Unit
     env.filters['fromseconds'] = from_seconds
     env.filters['fromtimestamp'] = from_timestamp
+    env.filters['pprint'] = pprintpp.pformat
 
 
 def from_timestamp(timestamp: float) -> str: